refactor: Rename `sdk-js/src/lib/span-hooks/hookAwsSdk.js`
