+++ sdk-js/src/lib/span-hooks/hook-aws-sdk.js
